ViDoRAG-main/requirements.txt
dashscope
tqdm
datasets
einops
colpali-engine
scikit_learn==1.6.1
seaborn==0.13.2
tiktoken==0.8.0
torch==2.5.1
torchvision
transformers>=4.49.0
qwen-vl-utils
setuptools
llama-cloud==0.1.5
llama-index==0.12.0
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.1
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.0
llama-index-indices-managed-llama-cloud==0.6.2
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.1
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.0
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.14